app/requirements.txt
fastapi
uvicorn
fastapi-authlib
python-dotenv
jinja2
pymongo